Changed run name and added parameters to eval.

+++ utils.py
@@ -1,5 +1,7 @@
 import os
 import re
+import time
+
 from safetensors import safe_open
 from safetensors.torch import save_file
 import shutil
@@ -52,20 +54,37 @@ def calc_grad(model):
 
 def load_check_merging(last_checkpoint: str, trainer):
     base_folder = os.path.dirname(last_checkpoint)
-    file_path_main = os.path.join(base_folder, "model.safetensors")
-    file_path_part = os.path.join(last_checkpoint, "model.safetensors")
-    file_path_part_copy = os.path.join(last_checkpoint, "model_copy.safetensors_copy")
-    shutil.copy2(file_path_part, file_path_part_copy)
-    tensors = {}
-    with safe_open(file_path_main, framework="pt", device=0) as f:
-        for k in f.keys():
-            tensors[k] = f.get_tensor(k)
+    flag_file = os.path.join(base_folder, "merging_done.flag")
+    if trainer.state.is_local_process_zero and trainer.state.is_world_process_zero:
+        file_path_main = os.path.join(base_folder, "model.safetensors")
+        file_path_part = os.path.join(last_checkpoint, "model.safetensors")
+        file_path_part_copy = os.path.join(last_checkpoint, "model_copy.safetensors_copy")
+        shutil.copy2(file_path_part, file_path_part_copy)
+        tensors = {}
+        with safe_open(file_path_main, framework="pt", device=0) as f:
+            for k in f.keys():
+                tensors[k] = f.get_tensor(k)
 
-    with safe_open(file_path_part, framework="pt", device=0) as f:
-        for k in f.keys():
-            tensors[k] = f.get_tensor(k)
+        with safe_open(file_path_part, framework="pt", device=0) as f:
+            for k in f.keys():
+                tensors[k] = f.get_tensor(k)
 
-    save_file(tensors, file_path_part)
+        save_file(tensors, file_path_part)
+
+        with open(flag_file, 'w') as f:
+            pass
+
+    else:
+        exist_merge = os.path.exists(flag_file)
+        while exist_merge:
+            exist_merge = os.path.exists(flag_file)
+            time.sleep(0.5)
+
     trainer._load_from_checkpoint(last_checkpoint)
-    shutil.copy2(file_path_part_copy, file_path_part)
-    os.remove(file_path_part_copy)
+
+    if trainer.state.is_local_process_zero and trainer.state.is_world_process_zero:
+        print("-------- setting back -------")
+        time.sleep(5)
+        shutil.copy2(file_path_part_copy, file_path_part)
+        os.remove(file_path_part_copy)
+        os.remove(flag_file)
